Auth refactor, update snapshots
