+++ lib/aggregation-area/hooks/use-create-aggregation-areas.ts
@@ -1,5 +1,6 @@
 import {useCallback} from 'react'
 
+import {useCurrentRegionId} from 'lib/hooks/use-current-region'
 import useUser from 'lib/hooks/use-user'
 
 import createAggregationAreas from '../mutations/create-aggregation-areas'
@@ -7,7 +8,8 @@ import createAggregationAreas from '../mutations/create-aggregation-areas'
 /**
  * Use create
  */
-export default function useCreateAggregationAreas(regionId: string) {
+export default function useCreateAggregationAreas() {
+  const regionId = useCurrentRegionId()
   const {user} = useUser()
   return useCallback(
     (formData: FormData) => createAggregationAreas(formData, regionId, user),
