--- lib/selectors/regional-analyses.js
@@ -1,4 +1,0 @@
-import get from 'lodash/get'
-export default function selectRegionalAnalyses(state) {
-  return get(state, 'regionalAnalyses.analyses', [])
-}
